Move Tax Model to \App\Models\Settings\Taxes\
